[Easy] remove unnecessary lines (#28)
Adresses #25

The final claim will always be from secondaryToken back to depositToken (so that we have depositTokens to refund). Thus it's unnecessary to compute buy and sell token in that method.

There is another if statement, which is unnecessary since the call inside can already never fail (it might be slightly more gas efficient in the first auction, but I believe having less branches is more valuable.