Require an Infura key for Rinkeby/Mainnet interaction (#83)
We have seen a number of errors with our DxMgnPool bot over the last night. These are likely related to not using a Infura access token for interactions (cf. https://blog.infura.io/infura-dashboard-transition-update-c670945a922a)

This PR makes it so that we use the INFURA_KEY environment variable as a token.